+++ config/src/lib.rs
@@ -3,7 +3,7 @@
 #![deny(elided_lifetimes_in_paths)]
 #![warn(missing_docs)]
 #![warn(clippy::pedantic)]
-#![allow(clippy::module_name_repetitions, clippy::must_use_candidate)]
+#![allow(clippy::missing_errors_doc, clippy::module_name_repetitions, clippy::must_use_candidate)]
 
 mod config;
 mod format;