+++ CHANGELOG.md
@@ -13,6 +13,13 @@ Possible log types:
 
 ### Unreleased
 
+- [added] Add Nordic's nRF9160 MCU support
+- [added] Add Segger J-Link probe support
+- [removed] Remove `probe.itm.encoding` configuration in `Cargo.toml`
+- [added] Add `gdb-mi` task in generated `Justfile`
+- [fixed] Display Rust's `core`/`alloc` sources inside a GDB session
+- [changed] `release` profile includes debug symbols by default
+
 ### v0.11.1 (2019-11-27)
 
 - [added] New `expand` task for generated `Justfile`s